[TESTS] Reduce*Transformation template functional test-class and builder
