+++ inference-engine/tests/functional/plugin/shared/src/low_precision_transformations/reduce_max_transformation.cpp
@@ -36,14 +36,12 @@ void ReduceMaxTransformation::SetUp() {
     ngraph::pass::low_precision::LayerTransformation::Params params;
     ReduceMaxTransformationParam param;;
     std::tie(netPrecision, inputShape, targetDevice, params, param) = GetParam();
-    std::string reduceType = "Max";
 
-    function = ngraph::builder::subgraph::ReduceFunction::getOriginal(
+    function = ngraph::builder::subgraph::ReduceFunction::getOriginal<ngraph::opset1::ReduceMax>(
         netPrecision,
         inputShape,
         param.fakeQuantize,
         param.constantValues,
-        reduceType,
         param.keepDims);
 }
