+++ inference-engine/tests/functional/plugin/shared/src/low_precision_transformations/reduce_min_transformation.cpp
@@ -36,14 +36,12 @@ void ReduceMinTransformation::SetUp() {
     ngraph::pass::low_precision::LayerTransformation::Params params;
     ReduceMinTransformationParam param;;
     std::tie(netPrecision, inputShape, targetDevice, params, param) = GetParam();
-    std::string reduceType = "Min";
 
-    function = ngraph::builder::subgraph::ReduceFunction::getOriginal(
+    function = ngraph::builder::subgraph::ReduceFunction::getOriginal<ngraph::opset1::ReduceMin>(
         netPrecision,
         inputShape,
         param.fakeQuantize,
         param.constantValues,
-        reduceType,
         param.keepDims);
 }
